refactor(examples): group the igloo demo so it reads at a glance
Fifteen flat files where six were decoration. The top level is now the API
demonstration and nothing else; icons/ and art/ hold the theme's own glyphs and
scenery, so a reader can follow the composition without stepping through them.

Renames the two files whose names were confusingly close: frost.ts was a hook
and frost-svg.tsx a component. Now useFrost.ts and icons/Frost.tsx, which says
which is which without opening either.

Also drops an unused import and makes the toolbar's vertical padding symmetric
— it was 2px top and 8px bottom, which centres every control against its
siblings but sits the whole row 3px above the bar's own centre.
