+++ examples/igloo/src/art/random.ts
