+++ examples/igloo/src/igloo.css
@@ -203,17 +203,19 @@
   }
 }
 
-.igloo-menubar,
-.igloo-toolbar {
+.igloo-menubar {
   padding: 2px 12px;
 }
 
+/* SYMMETRIC vertical padding. It was 2px top and 8px bottom, which centres every control
+ * against its siblings but sits the whole row 3px above the bar's own centre — enough that
+ * the gap under the toolbar reads as wider than the gap over it. */
 .igloo-toolbar {
   display: flex;
   align-items: center;
   gap: 2px;
   flex-wrap: wrap;
-  padding-bottom: 8px;
+  padding: 6px 12px;
 }
 
 .igloo-toolbar__spacer {
